requirements.txt
bs4
datetime
google-search-results
googlemaps
nsfsearch
pypatent
scholarly
selenium
requests
bs4
beautifulsoup4
lxml
html5
uspto-opendata-python
collection
clinical_trials
pandas_read_xml
nested-lookup
